renaming text.rs module
