+++ src/body.rs
@@ -10,7 +10,7 @@ use crate::{
     cmap::ToUnicodeCMap,
     filters::flate_decode,
     object::{Array, Dictionary, Name, Number, Object},
-    text,
+    content,
     xref::XrefTable,
     Extract,
 };
@@ -442,7 +442,7 @@ impl Page {
             .get_resources()
             .font
             .expect("Missing font in current page resources");
-        text::TextContent::from(self.extract_stream().as_bytes()).get_text(fontmap)
+        content::TextContent::from(self.extract_stream().as_bytes()).get_text(fontmap)
     }
 
     pub fn extract_stream(&self) -> String {
