+++ src/lib.rs
@@ -7,7 +7,7 @@ pub mod cmap;
 pub mod filters;
 pub mod info;
 pub mod object;
-pub mod text;
+pub mod content;
 pub mod tokenizer;
 pub mod trailer;
 pub mod xref;
